requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dotenv==1.0.0
requests==2.31.0
scapy==2.5.0
urllib3==2.0.5
